requirements.txt
# Core dependencies
numpy>=1.21.0
pandas>=1.3.0
scipy>=1.7.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Change point detection
torch>=1.9.0
ruptures>=1.1.0
bayesian_changepoint_detection @ git+https://github.com/estcarisimo/bayesian_changepoint_detection.git

# Data input/output
influxdb-client>=1.30.0
requests>=2.25.0

# CLI and configuration
click>=8.0.0
rich>=12.0.0
typer>=0.9.0
pyyaml>=6.0

# API server
fastapi>=0.104.0
uvicorn>=0.24.0

# Development and testing
pytest>=6.2.0
pytest-cov>=3.0.0
black>=22.0.0
isort>=5.10.0
flake8>=4.0.0
mypy>=0.910

# Optional visualization
matplotlib>=3.5.0
plotly>=5.0.0

# Jupyter support (optional)
ipykernel>=6.0.0
jupyter>=1.0.0